Updated Lab 5 title and indentation of code snippets
In Lab 5: Access and Load Partitioned Object Storage Data, Rick changed the title from "Query External Data" to "Access and Load Partitioned Object Storage Data", to match the title in the navigation tree. Rick also changed the indentation of 3 or 4 code snippets in Tasks 2 and 3, so they appear correct in browser.