Workshop ID#142 Add Lab 8: Create a Clone Database, update Lab 7: Enable Disaster Recovery
Workshop ID#142
For Lab 7, updated screenshots and text to show how to enable Autonomous Data Guard, and created a new Task 4 on the new recovery type, Backup-Based Disaster Recovery.

For Lab 8, added this new lab introducing how to create a clone database from the current primary database and from a backup, and how to create a long-term backup.