WMS ID : 553 - Added task to scale down OCPU allocation to reduce over allocation of OCPU and match with rest of the labs (#201)
* Updated pages for ECPU to OCPU conversion.

As suggested by David, I have updated the copied adb-provision-conditional.md file into adb-provision-conditional-ecpu.md file and added the explanation with images to change from ECPU to OCPU option.  Also copies the livelabs-indepth folder under  shared/workshops folder and updated the manifest.json file to point to  adb-provision-conditional-ecpu.md.  These changes are to accommodate recent adoption of ECPU metric on oraclepartnersas tenancy.

* Updated the Scale down OCPU and Storage Task for PTS version only

As discussed with David Start, have added a Task to scale down OCPU allocation to match the rest of the lab exercises and prevent over-allocation of OCPU, added markdown files with pts in the filename to indicate new files, and updated the link in the manifest file under livelabs-indepth-ecpu folder.