Workshop ID#142 updates for ECPU replacing OCPU plus other UI changes
These updates are primarily related to the release of the ECPU compute model as the default model instead of OCPU. Most of the updates are to the provisioning lab and the auto scaling labs. Other labs were updated to reflect the renaming of the "Manage scaling" button to "Manage resource allocation".