Workshop ID#141 Load and Analyze fixed advice on naming database in adb-provision.md
Workshop ID#141 Load and Analyze (actually not used by that or any of my workshops; other authors are using the adb-provision.md file, while I use adb-provision-conditional.md version). Chris Hoina is using this version.
I fixed the step on naming the display name and database name, to not use blank spaces or underscores in display name or database name.